Move unit test to a separate file
